+++ src/main/java/de/dm/intellij/liferay/schema/LiferayJsonSchemaProviderFactory.java
@@ -15,6 +15,7 @@ public class LiferayJsonSchemaProviderFactory implements JsonSchemaProviderFacto
     public List<JsonSchemaFileProvider> getProviders(@NotNull Project project) {
         return Arrays.asList(
             new LiferayJournalStructureJsonSchemaFileProvider(project),
+            new LiferayJournalStructureJsonSchema_2_0_FileProvider(project),
             new LiferayAssetsJsonSchemaFileProvider(project),
             new LiferaySitemapJsonSchemaFileProvider(project),
             new LiferayFrontendTokenDefinitionJsonSchemaFileProvider(project)