+++ changelog.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Feature: Language.properties references for label attributes in frontend-token-definition.json  
 - Feature: Clay SCSS Variables documentation
 - Feature: Added support for portlet-display-templates.xml (XML Schema and file references)
+- Feature: Added support for new journal structure json schema (for Liferay 7.4)
 - Bugfix: Fixed detecting the bundles directory from the bundle-support maven plugin
   
 ## 0.0.11